Fix compression calculation again.
Fix bad cast in 3a604563c, and update pct to 64-bits to prevent
truncation of intermediate value (pct * 100).

+++ src/common/file_bcast.c
@@ -373,7 +373,7 @@ static int _bcast_file(struct bcast_parameters *params)
 	xfree(buffer);
 
 	if (size_uncompressed && params->compress != 0) {
-		int32_t pct = (int32_t)(size_uncompressed - size_compressed);
+		int64_t pct = (int64_t) size_uncompressed - size_compressed;
 		/* Dividing a negative by a positive in C99 results in
 		 * "truncation towards zero" which gives unexpected values for
 		 * pct. This construct avoids that problem.
@@ -381,7 +381,7 @@ static int _bcast_file(struct bcast_parameters *params)
 		pct = (pct>=0) ? pct * 100 / size_uncompressed
 			       : - (-pct * 100 / size_uncompressed);
 		verbose("File compressed from %u to %u (%d percent) in %u usec",
-			size_uncompressed, size_compressed, pct,
+			size_uncompressed, size_compressed, (int) pct,
 			time_compression);
 	}
